config/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
dotenv==0.9.9
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
